+++ signin.html
@@ -38,14 +38,14 @@
             <div class="sign-input-box-include-error sign-password">
               <input class="sign-input" type="password" id="passwordInput" />
               <p id="error-message-password" class="display-none"></p>
-              <button class="eye-button" type="button">
+              <button class="eye-button" type="button" id="eyeBtn">
                 <img src="./images/eye-off.svg" id="offEye" />
                 <img src="./images/eye-on.svg" id="onEye" class="display-none" />
               </button>
             </div>
           </div>
         </div>
-        <button class="cta" type="submit" id="loginBtn">로그인</button>
+        <button class="cta" type="submit" id="submitBtn">로그인</button>
       </form>
       <div class="sns-box">
         <span class="sns-text">소셜 로그인</span>
@@ -59,6 +59,6 @@
         </div>
       </div>
     </div>
-    <script src="./src/JS/sign.js"></script>
+    <script type="module" src="./src/JS/signin.js"></script>
   </body>
-</html>
+  </html>
