+++ src/JS/sign.js
@@ -1,15 +1,16 @@
+// sign 공통변수
 const signForm = document.querySelector('#signForm');
 const errorMessageEmail = document.querySelector('#error-message-email');
 const errorMessagePassword = document.querySelector('#error-message-password');
 const emailInput = document.querySelector('#emailInput');
 const passwordInput = document.querySelector('#passwordInput');
-const loginBtn = document.querySelector('#loginBtn');
-const eyeBtn = document.querySelector('.eye-button');
+const submitBtn = document.querySelector('#submitBtn');
+const eyeBtn = document.querySelector('#eyeBtn');
 const eyeOff = document.querySelector('#offEye');
 const eyeOn = document.querySelector('#onEye');
 
 function checkValidationEmail(email_address){     
-	email_regex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/i;
+	let email_regex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/i;
 	if(!email_regex.test(email_address)){ 
 		return false; 
 	} else {
@@ -51,32 +52,6 @@ function checkPassword(){
   }
 }; // 비밀번호 유효성 체크
 
-function loginCheck(){
-  if(checkEmail() && checkPassword()){ // 먼저 이메일과 비밀번호의 유효성 확인
-    if(emailInput.value === 'test@codeit.com' && passwordInput.value === "codeit101"){
-      return true;
-    } else if (emailInput.value !== 'test@codeit.com' && passwordInput.value !== "codeit101"){
-      errorMsgAdd(emailInput, errorMessageEmail, '이메일을 확인해 주세요');
-      errorMsgAdd(passwordInput, errorMessagePassword, "비밀번호를 확인해 주세요");
-      return false;
-    } else if (emailInput.value !== 'test@codeit.com'){
-      errorMsgAdd(emailInput, errorMessageEmail, '이메일을 확인해 주세요');
-      return false;
-    } else if (passwordInput.value !== "codeit101"){
-      errorMsgAdd(passwordInput, errorMessagePassword, "비밀번호를 확인해 주세요");
-      return false;
-    };
-  };
-}; // 허락된 계정인지 확인
-
-function tryLogin(e){
-  e.preventDefault();
-  if(loginCheck()){
-    let link = '/folder.html';
-    window.location.href = link;
-  };
-}; // 로그인 시도
-
 function toggleEyeBtn(){
   if(passwordInput.getAttribute('type') === 'password'){
     passwordInput.setAttribute('type', 'text'); // 비밀번호 보이기
@@ -89,13 +64,4 @@ function toggleEyeBtn(){
   }
 }; // 비밀번호 보이기/가리기
 
-// 이벤트 관리
-emailInput.addEventListener('focusout', checkEmail);
-passwordInput.addEventListener('focusout', checkPassword);
-loginBtn.addEventListener('click', tryLogin);
-signForm.addEventListener('keydown', function(event){
-  if (event.key === 'Enter'){
-    tryLogin(e);
-  };
-});
-eyeBtn.addEventListener('click', toggleEyeBtn);
+export { signForm, errorMessageEmail, errorMessagePassword, emailInput, passwordInput, submitBtn, eyeBtn, eyeOff, eyeOn, checkValidationEmail, errorMsgAdd, errorMsgRemove, checkEmail, checkPassword, toggleEyeBtn };
